Fix bevy_ui compile error
