+++ crates/bevy_ui/src/picking_backend.rs
@@ -23,7 +23,7 @@
 #![allow(clippy::too_many_arguments)]
 #![deny(missing_docs)]
 
-use crate::{prelude::*, UiStack};
+use crate::{pick_rounded_rect, prelude::*, UiStack};
 use bevy_app::prelude::*;
 use bevy_ecs::{prelude::*, query::QueryData};
 use bevy_math::Vec2;
@@ -216,27 +216,4 @@ pub fn ui_picking(
 
         output.send(PointerHits::new(*pointer, picks, order));
     }
-}
-
-// Returns true if `point` (relative to the rectangle's center) is within the bounds of a rounded rectangle with
-// the given size and border radius.
-//
-// Matches the sdf function in `ui.wgsl` that is used by the UI renderer to draw rounded rectangles.
-pub(crate) fn pick_rounded_rect(
-    point: Vec2,
-    size: Vec2,
-    border_radius: ResolvedBorderRadius,
-) -> bool {
-    let s = point.signum();
-    let r = (border_radius.top_left * (1. - s.x) * (1. - s.y)
-        + border_radius.top_right * (1. + s.x) * (1. - s.y)
-        + border_radius.bottom_right * (1. + s.x) * (1. + s.y)
-        + border_radius.bottom_left * (1. - s.x) * (1. + s.y))
-        / 4.;
-
-    let corner_to_point = point.abs() - 0.5 * size;
-    let q = corner_to_point + r;
-    let l = q.max(Vec2::ZERO).length();
-    let m = q.max_element().min(0.);
-    l + m - r < 0.
 }
